pyproject.toml
[project]
name = "jyotish-backend"
version = "0.1.0"
description = "High-precision astrology backend system"
authors = [{name = "Antigravity", email = "antigravity@google.com"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pyswisseph>=2.10.3.2",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "redis>=5.0.1",
    "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.26.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "mypy>=1.8.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
